AB#59699 Default fields are breaking schema generation
