+++ src/utils/form/index.ts
@@ -13,3 +13,4 @@ export * from './getNextId';
 export * from './getDisplayText';
 export * from './checkRecordValidation';
 export * from './getAccessibleFields';
+export * from './checkFieldValueName';
